+++ src/promptflow-core/promptflow/core/_serving/flow_invoker.py
@@ -10,6 +10,7 @@ from promptflow._utils.dataclass_serializer import convert_eager_flow_output_to_
 from promptflow._utils.flow_utils import dump_flow_result, is_executable_chat_flow
 from promptflow._utils.logger_utils import LoggerFactory
 from promptflow._utils.multimedia_utils import MultimediaProcessor
+from promptflow.contracts.run_info import Status
 from promptflow.core._connection import _Connection
 from promptflow.core._connection_provider._connection_provider import ConnectionProvider
 from promptflow.core._flow import AbstractFlowBase
@@ -116,6 +117,14 @@ class FlowInvoker:
             connections_to_ignore.extend(self.connections_name_overrides.keys())
             self.logger.debug(f"Flow invoker connections name overrides: {self.connections_name_overrides.keys()}")
             self.logger.debug(f"Ignoring connections: {connections_to_ignore}")
+            if not connection_provider:
+                # If user not pass in connection provider string, get from environment variable.
+                connection_provider = ConnectionProvider.get_instance(credential=self._credential)
+            else:
+                # Else, init from the string to parse the provider config.
+                connection_provider = ConnectionProvider.init_from_provider_config(
+                    connection_provider, credential=self._credential
+                )
             # Note: The connection here could be local or workspace, depends on the connection.provider in pf.yaml.
             connections = self.resolve_connections(
                 # use os.environ to override flow definition's connection since
@@ -123,7 +132,7 @@ class FlowInvoker:
                 connection_names=self.flow.get_connection_names(
                     environment_variables_overrides=os.environ,
                 ),
-                provider=ConnectionProvider.init_from_provider_config(connection_provider, credential=self._credential),
+                provider=connection_provider,
                 connections_to_ignore=connections_to_ignore,
                 # fetch connections with name override
                 connections_to_add=list(self.connections_name_overrides.values()),
@@ -172,6 +181,7 @@ class FlowInvoker:
             raise_ex=self.raise_ex,
             storage=storage,
             init_kwargs=self._init_kwargs,
+            env_exporter_setup=False,
         )
         self.executor.enable_streaming_for_llm_flow(self.streaming)
         self.logger.info("Promptflow executor initiated successfully.")
@@ -214,8 +224,11 @@ class FlowInvoker:
             returned_non_dict_output = False
         resolved_outputs = self._convert_multimedia_data_to_base64(output_dict)
         self._dump_invoke_result(result)
-        log_outputs = "<REDACTED>" if disable_input_output_logging else result.output
-        self.logger.info(f"Flow run result: {log_outputs}")
+        if result.run_info.status != Status.Completed:
+            self.logger.error(f"Flow run failed with error: {result.run_info.error}")
+        else:
+            log_outputs = "<REDACTED>" if disable_input_output_logging else result.output
+            self.logger.info(f"Flow run result: {log_outputs}")
         if not self.raise_ex:
             # If raise_ex is False, we will return the trace flow & node run info.
             return FlowResult(
@@ -258,8 +271,11 @@ class AsyncFlowInvoker(FlowInvoker):
             returned_non_dict_output = False
         resolved_outputs = self._convert_multimedia_data_to_base64(output_dict)
         self._dump_invoke_result(result)
-        log_outputs = "<REDACTED>" if disable_input_output_logging else result.output
-        self.logger.info(f"Flow run result: {log_outputs}")
+        if result.run_info.status != Status.Completed:
+            self.logger.error(f"Flow run failed with error: {result.run_info.error}")
+        else:
+            log_outputs = "<REDACTED>" if disable_input_output_logging else result.output
+            self.logger.info(f"Flow run result: {log_outputs}")
         if not self.raise_ex:
             # If raise_ex is False, we will return the trace flow & node run info.
             return FlowResult(