+++ src/promptflow-tracing/promptflow/tracing/_trace.py
@@ -2,6 +2,7 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # ---------------------------------------------------------
 
+import asyncio
 import contextlib
 import functools
 import inspect
@@ -30,10 +31,10 @@ IS_LEGACY_OPENAI = version("openai").startswith("0.")
 
 
 @contextlib.contextmanager
-def _record_keyboard_interrupt_to_span(span: Span):
+def _record_cancellation_exceptions_to_span(span: Span):
     try:
         yield
-    except KeyboardInterrupt as ex:
+    except (KeyboardInterrupt, asyncio.CancelledError) as ex:
         if span.is_recording():
             span.record_exception(ex)
             span.set_status(StatusCode.ERROR, "Execution cancelled.")
@@ -181,7 +182,7 @@ def traced_generator(original_span: ReadableSpan, inputs, generator):
     with otel_tracer.start_as_current_span(
         f"Iterated({original_span.name})",
         links=[link],
-    ) as span, _record_keyboard_interrupt_to_span(span):
+    ) as span, _record_cancellation_exceptions_to_span(span):
         enrich_span_with_original_attributes(span, original_span.attributes)
         # Enrich the new span with input before generator iteration to prevent loss of input information.
         # The input is as an event within this span.
@@ -205,7 +206,7 @@ async def traced_async_generator(original_span: ReadableSpan, inputs, generator)
     with otel_tracer.start_as_current_span(
         f"Iterated({original_span.name})",
         links=[link],
-    ) as span, _record_keyboard_interrupt_to_span(span):
+    ) as span, _record_cancellation_exceptions_to_span(span):
         enrich_span_with_original_attributes(span, original_span.attributes)
         # Enrich the new span with input before generator iteration to prevent loss of input information.
         # The input is as an event within this span.
@@ -382,7 +383,7 @@ def _traced_async(
         span_name = get_node_name_from_context(used_for_span_name=True) or trace.name
         # need to get everytime to ensure tracer is latest
         otel_tracer = otel_trace.get_tracer("promptflow")
-        with otel_tracer.start_as_current_span(span_name) as span, _record_keyboard_interrupt_to_span(span):
+        with otel_tracer.start_as_current_span(span_name) as span, _record_cancellation_exceptions_to_span(span):
             # Store otel trace id in context for correlation
             OperationContext.get_instance()["otel_trace_id"] = f"0x{format_trace_id(span.get_span_context().trace_id)}"
             enrich_span_with_trace(span, trace)
@@ -448,7 +449,7 @@ def _traced_sync(
         span_name = get_node_name_from_context(used_for_span_name=True) or trace.name
         # need to get everytime to ensure tracer is latest
         otel_tracer = otel_trace.get_tracer("promptflow")
-        with otel_tracer.start_as_current_span(span_name) as span, _record_keyboard_interrupt_to_span(span):
+        with otel_tracer.start_as_current_span(span_name) as span, _record_cancellation_exceptions_to_span(span):
             # Store otel trace id in context for correlation
             OperationContext.get_instance()["otel_trace_id"] = f"0x{format_trace_id(span.get_span_context().trace_id)}"
             enrich_span_with_trace(span, trace)