+++ src/promptflow/tests/executor/e2etests/test_executor_execution_failures.py
@@ -1,12 +1,13 @@
+import re
+import sys
+
 import pytest
 
+from promptflow._core._errors import ToolExecutionError
 from promptflow.contracts.run_info import Status
 from promptflow.executor import FlowExecutor
-from promptflow._core._errors import ToolExecutionError
 
-from ..utils import (
-    get_yaml_file,
-)
+from ..utils import get_yaml_file
 
 SAMPLE_FLOW = "web_classification_no_variants"
 SAMPLE_EVAL_FLOW = "classification_accuracy_evaluation"
@@ -26,6 +27,7 @@ The above exception was the direct cause of the following exception:
 Traceback (most recent call last):
 in wrapped
     output = func(*args, **kwargs)
+             ^^^^^^^^^^^^^^^^^^^^^
 sync_fail.py", line 13, in raise_an_exception
     raise Exception(f"In tool raise_an_exception: {s}") from e
 Exception: In tool raise_an_exception: dummy_input
@@ -44,6 +46,7 @@ The above exception was the direct cause of the following exception:
 Traceback (most recent call last):
 in wrapped
     output = await func(*args, **kwargs)
+             ^^^^^^^^^^^^^^^^^^^^^^^^^^^
 in raise_an_exception_async
     raise Exception(f"In tool raise_an_exception_async: {s}") from e
 Exception: In tool raise_an_exception_async: dummy_input
@@ -51,6 +54,11 @@ Exception: In tool raise_an_exception_async: dummy_input
         "\n"
     ),
 }
+
+if sys.version_info < (3, 11):
+    # Python 3.11 on Mac has an extra line of ^^^^^ to point on the function raising the exception
+    for key in expected_stack_traces:
+        expected_stack_traces[key] = [line for line in expected_stack_traces[key] if re.match(r"^\s+\^+", line) is None]
 
 
 @pytest.mark.e2etest
@@ -81,7 +89,9 @@ class TestExecutorFailures:
         #  Make sure the stack trace is as expected
         stacktrace = user_error_info["traceback"].split("\n")
         expected_stack_trace = expected_stack_traces[flow_folder]
-        assert len(stacktrace) == len(expected_stack_trace)
+        if len(stacktrace) != len(expected_stack_trace):
+            # actually we should fail now; adding this to make sure we can see the difference
+            assert stacktrace == expected_stack_trace
         for expected_item, actual_item in zip(expected_stack_trace, stacktrace):
             assert expected_item in actual_item
 
@@ -112,7 +122,9 @@ class TestExecutorFailures:
         #  Make sure the stack trace is as expected
         stacktrace = user_error_info["traceback"].split("\n")
         expected_stack_trace = expected_stack_traces[flow_folder]
-        assert len(stacktrace) == len(expected_stack_trace)
+        if len(stacktrace) != len(expected_stack_trace):
+            # actually we should fail now; adding this to make sure we can see the difference
+            assert stacktrace == expected_stack_trace
         for expected_item, actual_item in zip(expected_stack_trace, stacktrace):
             assert expected_item in actual_item