+++ src/promptflow/tests/test_configs/eager_flows/dummy_callable_class_with_connection/simple_callable_with_connection.py
@@ -1,0 +1,10 @@
+from promptflow.connections import AzureOpenAIConnection
+
+
+class MyFlow:
+    def __init__(self, connection: AzureOpenAIConnection):
+        self._connection = connection
+
+    def __call__(self):
+        assert isinstance(self._connection, AzureOpenAIConnection)
+        return "Dummy output"