+++ src/promptflow-devkit/CHANGELOG.md
@@ -1,6 +1,16 @@
 # promptflow-devkit package
 
-## v1.10.0 (Upcoming)
+## v1.11.0 (Upcoming)
+
+### Improvements
+- Interactive browser credential is excluded by default when using Azure AI connections, user could set `PF_NO_INTERACTIVE_LOGIN=False` to enable it.
+- Visualize flex flow run(s) switches to trace UI page.
+
+### Bugs Fixed
+- Fix the issue that import error will be raised after downgrading promptflow from >=1.10.0 to <1.8.0.
+- Fix the issue that `pf flow serve` is broken with exception `NotADirectoryError`.
+
+## v1.10.0 (2024.04.26)
 
 ### Features Added
 - Expose --ui to trigger a chat window, reach [here](https://microsoft.github.io/promptflow/reference/pf-command-reference.html#pf-flow-test) for more details.